+++ R/globals.R
@@ -11,6 +11,11 @@ if (getRversion() >= "2.15.1") {
     "end", "i", "j", "group_id", "x_tot", "x_1", "x_2",
 
 
-    "ID", "type", "gene_id", "attribute", "gene_name", "seqid"
+    "ID", "type", "gene_id", "attribute", "gene_name", "seqid",
+
+    # plot_exclusive_junctions
+    "transcript_id", "transcript_name", "exon_number",
+    "j_start", "j_end", "exclusive", "targeted",
+    "n_tx", "has_excl", "y", "y_label", "xend", "attr", "seqname"
   ))
 }